requirements.txt
# local package
-e .

# runtime requirements
python-dateutil>=2.8.0; python_version == "3.11.0"
pandas>=2.0.0
numpy>=2.0.0
scikit-learn>=1.8.0
fastapi>=0.135.1
uvicorn>=0.42.0
kagglehub[pandas-datasets]>=1.0.0
pydantic>=2.12.5
dvc>=3.67.0
ruamel.yaml>=0.17.0
joblib>=1.0.0

# tooling requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
